feat: split relayer polling and server
