+++ packages/agents/relayer/package.json
@@ -5,7 +5,8 @@
   "author": "Connext",
   "license": "MIT",
   "scripts": {
-    "start": "node dist/index.js",
+    "start": "node dist/relayer.js",
+    "server": "node dist/server.js",
     "docker:redis:up": "docker run -it --rm --name redis -p 6379:6379 redis",
     "lint": "eslint ./src --ext .ts --env node",
     "test": "nyc ts-mocha --check-leaks --exit --timeout 120000 'test/**/*.spec.ts'",
