+++ packages/agents/relayer/src/index.ts
@@ -1,6 +1,3 @@
-import { makeRelayer } from "./relayer";
-
-export { RelayerConfig, ChainConfig } from "./lib/entities";
-export { makeRelayer };
+import { makeRelayer } from "./make";
 
 makeRelayer();
