+++ packages/agents/relayer/src/server.ts
@@ -1,0 +1,5 @@
+import { makeServer } from "./make";
+
+export { makeServer };
+
+makeServer();
